runge-kutta 4 order

+++ Oz.Simulation/Oz.SimulationLib/Components/MassComponent.cs
@@ -12,6 +12,6 @@ public class MassComponent : SimComponent
     {
     }
 
-    public void SetMass(int mass) =>
+    public void SetMass(double mass) =>
         Mass = mass;
 }
